🚚 `ProfileRepository` moved into `profile/domain` module (#9)
