+++ auth/domain/build.gradle
@@ -2,6 +2,7 @@ apply from: "$rootDir/plugins/kotlin-jvm.gradle"
 
 dependencies {
   implementation project(":core:domain")
+  implementation project(":profile:domain")
 
   implementation(libs.kotlinx.coroutines.core)
 }
