+++ auth/domain/src/main/java/app/tasky/auth/domain/usecase/auth/SignUpUseCase.kt
@@ -21,8 +21,8 @@ import app.tasky.auth.domain.repository.SignUpResult
 import app.tasky.auth.domain.repository.SignUpResult.Failure
 import app.tasky.auth.domain.usecase.validation.SignUpValidationUseCase
 import app.tasky.auth.domain.util.EmailAlreadyInUseException
-import app.tasky.core.domain.repository.ProfileRepository
 import app.tasky.core.domain.util.NoNetworkConnectionException
+import app.tasky.profile.domain.repository.ProfileRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.flow
